moved signature utility methods to PackageHelper
