+++ common/src/main/java/com/microsoft/identity/common/internal/broker/PackageHelper.java
@@ -30,12 +30,14 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.PackageManager.NameNotFoundException;
 import android.content.pm.Signature;
+import android.os.Build;
 import android.util.Base64;
+
+import androidx.annotation.NonNull;
 
 import com.microsoft.identity.common.adal.internal.AuthenticationConstants;
 import com.microsoft.identity.common.adal.internal.util.StringExtensions;
 import com.microsoft.identity.common.internal.logging.Logger;
-import com.microsoft.identity.common.internal.util.SignUtil;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -67,8 +69,8 @@ public class PackageHelper {
      */
     public String getCurrentSignatureForPackage(final String packageName) {
         try {
-            final PackageInfo info = SignUtil.getPackageInfo(mPackageManager, packageName);
-            final Signature [] signatures = SignUtil.getSignatures(info);
+            final PackageInfo info = getPackageInfo(mPackageManager, packageName);
+            final Signature [] signatures = getSignatures(info);
             if (signatures != null && signatures.length > 0) {
                 final Signature signature = signatures[0];
                 MessageDigest md = MessageDigest.getInstance("SHA");
@@ -156,5 +158,54 @@ public class PackageHelper {
             }
         }
         return "";
+    }
+
+    /**
+     * Helper method to get signatures in a back-compatible way
+     *
+     * @param packageInfo A packageInfo instance with the flag PackageManager.GET_SIGNING_CERTIFICATES/PackageManager.GET_SIGNATURES set
+     * @return Signature[] or null
+     */
+    public static Signature[] getSignatures(final PackageInfo packageInfo) {
+        if (packageInfo == null) {
+            return null;
+        }
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            if (packageInfo.signingInfo == null) {
+                return null;
+            }
+            if (packageInfo.signingInfo.hasMultipleSigners()) {
+                return packageInfo.signingInfo.getApkContentsSigners();
+            } else {
+                return packageInfo.signingInfo.getSigningCertificateHistory();
+            }
+        }
+
+        return packageInfo.signatures;
+    }
+
+    public static int getPackageManagerFlag() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            return PackageManager.GET_SIGNING_CERTIFICATES;
+        }
+
+        return PackageManager.GET_SIGNATURES;
+    }
+
+    public static PackageInfo getPackageInfo(@NonNull PackageManager packageManager, @NonNull String packageName) throws PackageManager.NameNotFoundException {
+        return packageManager.getPackageInfo(packageName, getPackageManagerFlag());
+    }
+
+    public static PackageInfo getPackageInfo(@NonNull Context context, @NonNull String packageName) throws PackageManager.NameNotFoundException {
+        return getPackageInfo(context.getPackageManager(), packageName);
+    }
+
+    public static PackageInfo getPackageInfo(@NonNull Context context) throws PackageManager.NameNotFoundException {
+        return getPackageInfo(context, context.getPackageName());
+    }
+
+    public static Signature[] getSignatures(@NonNull Context context) throws PackageManager.NameNotFoundException {
+        return getSignatures(getPackageInfo(context));
     }
 }
